move targets plugin manager to plugin dir
